refactor: remove customFilter to util
